requirements.txt
f5-tts
soundfile
ruaccent

streamlit

matplotlib
numpy
pandas
seaborn

python-dotenv

torchvision